+++ aa_forum/static/aa_forum/ckeditor/plugins/youtube/lang/cs.js
@@ -1,3 +1,5 @@
+/* global CKEDITOR */
+
 CKEDITOR.plugins.setLang('youtube', 'cs', {
 	button : 'Vložit video YouTube',
 	title : 'Vložit video YouTube',